# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spacca POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/active-orders:
    get:
      operationId: getActiveOrders
      tags: [dashboard]
      summary: Get active orders for various screens
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, paid, in_progress, ready]
        - in: query
          name: branchId
          schema:
            type: integer
      responses:
        "200":
          description: Active orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDetail"
...
  /dashboard/sales-by-category:
    get:
      operationId: getSalesByCategory
      tags: [dashboard]
      summary: Get sales grouped by drink category
      parameters:
        - in: query
          name: days
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySales"
  /dashboard/top-drinks:
    get:
      operationId: getTopDrinks
      tags: [dashboard]
      summary: Get top selling drinks
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: days
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Top drinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopDrink"
...
components:
  schemas:
...
    OrderItemCustomization:
      type: object
      properties:
        id:
          type: integer
        orderItemId:
          type: integer
        ingredientId:
          type: ["integer", "null"]
        optionId:
          type: ["integer", "null"]
        typeVolumeId:
          type: ["integer", "null"]
        consumedQty:
          type: number
        producedQty:
          type: number
        addedCost:
          type: number
        slotLabel:
          type: string
        optionLabel:
          type: string
        baristaSortOrder:
          type: ["integer", "null"]
      required:
        - id
        - orderItemId
        - consumedQty
        - addedCost
        - slotLabel
        - optionLabel
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        drinkId:
          type: integer
        drinkName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        lineTotal:
          type: number
        specialNotes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, ready]
        readyAt:
          type: ["string", "null"]
        customizations:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemCustomization"
      required:
        - id
        - orderId
        - drinkId
        - drinkName
        - quantity
        - unitPrice
        - lineTotal
        - specialNotes
        - status
        - readyAt
        - customizations
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        baristaId:
          type: integer
        baristaName:
          type: string
        status:
          type: string
          enum: [pending, paid, in_progress, ready, completed, cancelled, refunded]
        customerName:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        discountId:
          type: ["integer", "null"]
        discountValue:
          type: ["number", "null"]
        discountType:
          type: ["string", "null"]
          enum: [percentage, fixed]
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, wallet]
        amountTendered:
          type: ["number", "null"]
        changeDue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        paidAt:
          type: ["string", "null"]
        readyAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
      required:
        - id
        - orderNumber
        - baristaId
        - baristaName
        - status
        - customerName
        - subtotal
        - discount
        - total
        - paymentMethod
        - amountTendered
        - changeDue
        - notes
        - createdAt
        - updatedAt
        - paidAt
        - readyAt
        - completedAt
        - cancelledAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required:
            - items
...
    CategorySales:
      type: object
      properties:
        category:
          type: string
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalDrinks:
          type: integer
      required:
        - category
        - totalOrders
        - totalRevenue
        - totalDrinks
    TopDrink:
      type: object
      properties:
        drinkId:
          type: integer
        drinkName:
          type: string
        category:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
      required:
        - drinkId
        - drinkName
        - category
        - totalSold
        - totalRevenue
...